make signup payment text minimum fifteen pounds
The board has decided to raise the minimum amount we ask for new sign ups to pay £15. Of course, people are still able to input a lower amount (even lower than £12.50) but this requires a longer discussion around code changes. I have however amended the text in the sign up to reflect that we are asking people this, so as to avoid having people join now and then ask them in 2, 3 (or n months) to raise it to £15. AFAIK emails have been sent to members paying below £15 (and who are active in the space) about raising their subscription.