Fix failed subscription payments being re-billed in later months
If a members subscription payment failed, we were transitoning it back to "due". This caused some confusing behaviour in the next months billing.

1. CheckMemberships would detect the user as having passed their expiry date, and suspend them.
2. BillMembers would find those previous "Due" charges and re-bill them, which also reactivated their account.

It's worth noting that the re-activation issue is new, following our recent bug fix which brought membership suspension back.

Prior to the re-activation issue, we had another strange sequencing issue: Such members would still be considered active, so CreateTodaysSubCharges (which runs a few days ahead of the billing charge date) would create the next month's subscription charge and mark it as "due" for such users. This meant that at month 1 after a payment failure, we attempt to charge the customer for 2 months worth of subscriptions. The next month, we charge 3. And then 4. Etc.

As GoCardless now charges us fees for each failed payment, these are incurring greater & greater fees each time we try to re-charge for a failed payment.

We could mark all subscription charges associated with failed payments as "cancelled" directly in the database to fix this issue. This code change is to provide additional protection in case there are timing issues (between CreateTodaysSubCharges and CheckMemberships) that introduces new "due" charges), or future code changes/bugs re-introduce this scenario.