Remove attempt to cancel GoCardless subscription on ban
This doesn't work, and it would be more appropriate to break this out into a separate action. In the case of temporary suspensions pending investigation, we wouldn't want to interrupt a member's billing set up as that only creates work for them if they become unbanned due to an investigation clearing them of any wrongdoing.