Remove paid inductions & refactor induction controller
We no longer require paid inductions for any pieces of equipment, and we're also going to be reducing the ways we take digital payments to make our membership system easier to maintain.

The wider refactor here helped break out the individual actions that can take place against inductions, and brings this area of the codebase more in line with Laravel development practises.