Require general induction before membership payments
This is only a frontend constraint. The backend will allow them to set up a subscription before doing the general induction if they manage to find the link to do so, but that shouldn't be linked anywhere. And they'll only be hampering themselves by doing that – paying for something they can't yet access.