Improve daily membership checks
Separate membership recovery out from being a side effects of the other processes, and remove CheckMemberships given it now duplicates what CheckSuspendedUsers does.

We shouldn't expect to see RecoverMemberships invoked in typical usage, but it's a useful stop gap in case other bugs appear in the system that cause data integrity issues. If we see it do anything, we should definitely investigate why it was necessary.